Add --client option to today/iteration
You can now specify a section `project_clients` in the config file,
which will associate a project with a client. This is kind of halfassed,
because Toggl has an API to ask about the clients, but I don't really
need that, so I'm not going to use it.

Syntax for the config file is:

```
[project_clients]
1234567 = "client 1"
76543   = "client 2"
```